refactor(sdf): rename to `sdf_information_schema.py` and `sdf_workspace.py`
